tolerate spaces in scaffold nouns
